+++ src/syntactic/syntactic_view.rs
@@ -314,7 +314,7 @@ impl SyntacticView {
             .unwrap_or(Color::AnsiValue(240));
         let match_bg = styles.preview_match.get_bg().unwrap_or(Color::AnsiValue(28));
         let code_width = area.width as usize - 1; // 1 char left for scrollbar
-        let scrollbar = area.scrollbar(self.scroll as i32, self.lines.len() as i32);
+        let scrollbar = area.scrollbar(self.scroll, self.lines.len());
         let scrollbar_fg = styles.scrollbar_thumb.get_fg()
             .or_else(|| styles.preview.get_fg())
             .unwrap_or(Color::White);
